requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
Pillow==10.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
